+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/CourierTest.java
@@ -8,12 +8,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.nsu.krasnyanskii.pizzeria.model.Order;
+import ru.nsu.krasnyanskii.pizzeria.view.ConsolePizzeriaView;
 import ru.nsu.krasnyanskii.pizzeria.view.PizzeriaView;
 import ru.nsu.krasnyanskii.pizzeria.workers.Courier;
 
 class CourierTest {
 
-    private final PizzeriaView view = new PizzeriaView();
+    private final PizzeriaView view = new ConsolePizzeriaView();
 
     @BeforeEach
     void resetCounter() {
