+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/PizzeriaIntegrationTest.java
@@ -7,6 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.nsu.krasnyanskii.pizzeria.model.Order;
+import ru.nsu.krasnyanskii.pizzeria.view.ConsolePizzeriaView;
 import ru.nsu.krasnyanskii.pizzeria.view.PizzeriaView;
 import ru.nsu.krasnyanskii.pizzeria.workers.Baker;
 import ru.nsu.krasnyanskii.pizzeria.workers.Courier;
@@ -14,7 +15,7 @@ import ru.nsu.krasnyanskii.pizzeria.workers.Courier;
 /** Integration test: verifies orders travel the full pipeline from queue to delivery. */
 class PizzeriaIntegrationTest {
 
-    private final PizzeriaView view = new PizzeriaView();
+    private final PizzeriaView view = new ConsolePizzeriaView();
 
     @BeforeEach
     void resetCounter() {
